Texture managerr introduction
- Added a texture manager to handle most of the texture tasks.
- Updated the game manager to properly assign all managers.
