+++ Assets/Scripts/TextureManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1bbbf21a2877f5e4082d50c076ad60af
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
